refactor: move core packages to root dir
