+++ api/server_test.go
@@ -9,8 +9,8 @@ import (
 	"github.com/gin-gonic/gin"
 	assertTest "github.com/stretchr/testify/assert"
 	"gopkg.in/h2non/gock.v1"
-	"gopkg.in/sundowndev/phoneinfoga.v2/pkg/config"
-	"gopkg.in/sundowndev/phoneinfoga.v2/pkg/scanners"
+	"gopkg.in/sundowndev/phoneinfoga.v2/config"
+	"gopkg.in/sundowndev/phoneinfoga.v2/scanners"
 )
 
 var r *gin.Engine
