+++ cmd/version.go
@@ -2,8 +2,8 @@ package cmd
 
 import (
 	"github.com/spf13/cobra"
-	"gopkg.in/sundowndev/phoneinfoga.v2/pkg/config"
-	"gopkg.in/sundowndev/phoneinfoga.v2/pkg/utils"
+	"gopkg.in/sundowndev/phoneinfoga.v2/config"
+	"gopkg.in/sundowndev/phoneinfoga.v2/utils"
 )
 
 func init() {
