+++ scanners/google_test.go
@@ -4,7 +4,7 @@ import (
 	"testing"
 
 	assertTest "github.com/stretchr/testify/assert"
-	"gopkg.in/sundowndev/phoneinfoga.v2/pkg/utils"
+	"gopkg.in/sundowndev/phoneinfoga.v2/utils"
 )
 
 func TestGoogleSearchScan(t *testing.T) {
