+++ scanners/numverify.go
